confirmation: Support expiry date for validating confirmation link.
We check expiry of confirmation link by comparing the time elapsed
since the day link was sent to predefined no. of days for
each confirmation link.

This adds support for specifiying expiry datetime for confirmation links,
we pass an extra argument, no. of days (validity_in_days), to the
`create_confirmation_link`  which is used for determining expiry_date of
the link which is then saved to confirmation object and is later used for
checking expiry of a confirmation link by comparing it to current date-time.

Part of #16359.